Fix : mainAPI 코드 수정
